chore: split api e2e tests to get runtimes under control, and move out of parent account
