+++ .vscode/settings.json
@@ -6,7 +6,7 @@
     "**/build": true,
     "**/coverage": true
   },
-  "editor.formatOnSave": true,
+  "editor.formatOnSave": false,
   "eslint.validate": ["javascript", "javascriptreact", "typescript", "typescriptreact"],
   "eslint.alwaysShowStatus": true,
   "eslint.format.enable": true,
@@ -14,7 +14,7 @@
   "eslint.packageManager": "yarn",
   "eslint.quiet": true,
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
+    "source.fixAll.eslint": false
   },
   "jest.enableInlineErrorMessages": true,
   "jest.showCoverageOnLoad": true,
